requirements.txt
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
json5==0.8.4
jsonschema==3.0.1
MarkupSafe==1.1.1
matplotlib==3.1.0
more-itertools==7.0.0
mpmath==1.1.0
msgpack==0.6.1
numpy==1.16.4
numpydoc==0.9.1
pandas==0.24.2
pandocfilters==1.4.2
pathlib2==2.3.4
patsy==0.5.1
pep8==1.7.1
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psutil==5.6.3
py==1.8.0
pycodestyle==2.5.0
ruamel-yaml==0.15.46
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.2.1
seaborn==0.9.0
statsmodels==0.10.0
virtualenv==16.7.4
webcolors==1.9.1
webencodings==0.5.1
Flask==1.1.1
Flask-RESTful==0.3.7